ecommerce/requirements.txt
# requirements.txt
pymongo==4.5
pydantic==2.3
requests==2.31
email-validator==2.0
typing-extensions==4.8
pillow==10.0.1